code restructure: relocate and rename settings files
